--- a/benches/rustfft.rs
+++ b/benches/rustfft.rs
@@ -17,19 +17,21 @@ fn bench_fft(b: &mut Bencher, len: usize) {
     b.iter(|| {fft.process(&mut signal, &mut spectrum);} );
 }
 
-// Powers of 2
-#[bench] fn complex_p2_00128(b: &mut Bencher) { bench_fft(b,    128); }
-#[bench] fn complex_p2_00512(b: &mut Bencher) { bench_fft(b,   512); }
-#[bench] fn complex_p2_02048(b: &mut Bencher) { bench_fft(b,  2048); }
-#[bench] fn complex_p2_08192(b: &mut Bencher) { bench_fft(b,  8192); }
-#[bench] fn complex_p2_32768(b: &mut Bencher) { bench_fft(b, 32768); }
 
 // Powers of 4
-#[bench] fn complex_p4_00064(b: &mut Bencher) { bench_fft(b,    64); }
-#[bench] fn complex_p4_00256(b: &mut Bencher) { bench_fft(b,   256); }
-#[bench] fn complex_p4_01024(b: &mut Bencher) { bench_fft(b,  1024); }
-#[bench] fn complex_p4_04096(b: &mut Bencher) { bench_fft(b,  4096); }
-#[bench] fn complex_p4_16384(b: &mut Bencher) { bench_fft(b, 16384); }
+#[bench] fn complex_p2_000064(b: &mut Bencher) { bench_fft(b,   64); }
+#[bench] fn complex_p2_000128(b: &mut Bencher) { bench_fft(b,   128); }
+#[bench] fn complex_p2_000256(b: &mut Bencher) { bench_fft(b,   256); }
+#[bench] fn complex_p2_000512(b: &mut Bencher) { bench_fft(b,   512); }
+#[bench] fn complex_p2_001024(b: &mut Bencher) { bench_fft(b,   1024); }
+#[bench] fn complex_p2_002048(b: &mut Bencher) { bench_fft(b,   2048); }
+#[bench] fn complex_p2_004096(b: &mut Bencher) { bench_fft(b,   4096); }
+#[bench] fn complex_p2_008192(b: &mut Bencher) { bench_fft(b,   8192); }
+#[bench] fn complex_p2_016384(b: &mut Bencher) { bench_fft(b,   16384); }
+#[bench] fn complex_p2_032768(b: &mut Bencher) { bench_fft(b,   32768); }
+#[bench] fn complex_p2_065536(b: &mut Bencher) { bench_fft(b,   65536); }
+#[bench] fn complex_p2_1048576(b: &mut Bencher) { bench_fft(b, 1048576); }
+#[bench] fn complex_p2_2097152(b: &mut Bencher) { bench_fft(b, 2097152); }
 
 // Powers of 7
 #[bench] fn complex_p7_00343(b: &mut Bencher) { bench_fft(b,   343); }