--- a/src/plan.rs
+++ b/src/plan.rs
@@ -10,9 +10,10 @@ use algorithm::butterflies::*;
 use math_utils;
 
 
-const MIN_RADIX4_BITS: u32 = 8; // minimum size to consider radix 4 an option is 2^8 = 256
-const BUTTERFLIES: [usize; 8] = [2, 3, 4, 5, 6, 7, 8, 16];
-const COMPOSITE_BUTTERFLIES: [usize; 4] = [4, 6, 8, 16];
+const MIN_RADIX4_BITS: u32 = 5; // smallest size to consider radix 4 an option is 2^5 = 32
+const MAX_RADIX4_BITS: u32 = 16; // largest size to consider radix 4 an option is 2^16 = 65536
+const BUTTERFLIES: [usize; 9] = [2, 3, 4, 5, 6, 7, 8, 16, 32];
+const COMPOSITE_BUTTERFLIES: [usize; 5] = [4, 6, 8, 16, 32];
 
 /// The FFT planner is used to make new FFT algorithm instances.
 ///
@@ -81,6 +82,7 @@ impl<T: FFTnum> FFTplanner<T> {
                 7 => Arc::new(Butterfly7::new(inverse)),
                 8 => Arc::new(Butterfly8::new(inverse)),
                 16 => Arc::new(Butterfly16::new(inverse)),
+                32 => Arc::new(Butterfly32::new(inverse)),
                 _ => panic!("Invalid butterfly size: {}", len),
             }
         ).clone()
@@ -93,7 +95,7 @@ impl<T: FFTnum> FFTplanner<T> {
             let result = if factors.len() == 1 || COMPOSITE_BUTTERFLIES.contains(&len) {
                 self.plan_fft_single_factor(len)
 
-            } else if len.trailing_zeros() >= MIN_RADIX4_BITS {
+            } else if len.trailing_zeros() <= MAX_RADIX4_BITS && len.trailing_zeros() >= MIN_RADIX4_BITS {
                 //the number of trailing zeroes in len is the number of `2` factors
                 //ie if len = 2048 * n, len.trailing_zeros() will equal 11 because 2^11 == 2048
 
@@ -189,6 +191,7 @@ impl<T: FFTnum> FFTplanner<T> {
             7 => Arc::new(butterflies::Butterfly7::new(self.inverse)) as Arc<FFT<T>>,
             8 => Arc::new(butterflies::Butterfly8::new(self.inverse)) as Arc<FFT<T>>,
             16 => Arc::new(butterflies::Butterfly16::new(self.inverse)) as Arc<FFT<T>>,
+            32 => Arc::new(butterflies::Butterfly32::new(self.inverse)) as Arc<FFT<T>>,
             _ => self.plan_prime(len),
         }
     }