This PR does 4 main things:
- Rewrites Butterfly16 to use a split--2-4-radix setup instead of mixed radix, for a 2x speedup, and adds a butterfly32 with the same design
- Rewrites Radix4 to use Butterfly8 and Butterfly16 as a base case instead of Butterfly2 and Butterfly. This improves the speed at small sizes
- Rewrites Radix4 twiddle factors so that instead of having a fixed list that it scans over with different strides, it groups the twiddles each layer needs into a block that is accessed with a stride of 1. This takes more memory, but gives a 10-20% improvement at larger sizes.
- Places an upper limit in the planner on which sizes will use radix 4. Above 2^16, it uses mixed radix instead. This has a huge effect when the problem doesn't fit into cache, because I write the mixed radix algorithm to be very cache friendly.

Benchmarks before:
test complex_p2_00000064       ... bench:         271 ns/iter (+/- 33)
test complex_p2_00000256       ... bench:       1,735 ns/iter (+/- 150)
test complex_p2_00001024       ... bench:       8,509 ns/iter (+/- 758)
test complex_p2_00004096       ... bench:      41,676 ns/iter (+/- 3,838)
test complex_p2_00016384       ... bench:     201,696 ns/iter (+/- 18,394)
test complex_p2_00065536       ... bench:   1,052,056 ns/iter (+/- 100,591)
test complex_p2_01048576       ... bench:  40,949,208 ns/iter (+/- 2,464,292)
test complex_p2_16777216       ... bench: 1,554,480,456 ns/iter (+/- 22,497,409)

Benchmarks after:
test complex_p2_00000064       ... bench:         265 ns/iter (+/- 45)
test complex_p2_00000256       ... bench:       1,390 ns/iter (+/- 139)
test complex_p2_00001024       ... bench:       7,013 ns/iter (+/- 608)
test complex_p2_00004096       ... bench:      35,085 ns/iter (+/- 2,444)
test complex_p2_00016384       ... bench:     173,429 ns/iter (+/- 15,180)
test complex_p2_00065536       ... bench:     812,941 ns/iter (+/- 73,863)
test complex_p2_01048576       ... bench:  20,780,458 ns/iter (+/- 1,263,067)
test complex_p2_16777216       ... bench: 409,238,648 ns/iter (+/- 13,505,602)

2^16 to 2^18 is the point where the whole thing stops fitting in L3 cache, and as far as i can tell, the higher you go, the more dramatic the improvement