@awelkie 2017-08-06
Looks good to me. It looks like the memory usage for Radix4's twiddle factors goes from `n` to `4n` where `n` is the length of the FFT. Is that correct? For what sizes did you see the 10%-20% speed gains with this change?