@ejmahler 2017-08-07
The memory usage goes from `n` to roughly `1.25n`

It uses n twiddle factors at the top level, n/4 twiddle factors at the next level, n/16 twiddle factors at the next level. etc. or n/x^2 in general. Using wolfram alpha to sum that up: https://www.wolframalpha.com/input/?i=sum(1+%2F+x+%5E+2+for+x%3D1+to+infinity)

You get ~1.64 * n. Then, 1/4 of these twiddle factors will always be 0, so you can omit them. 1.64*.75 = ~1.23