@ejmahler 2017-08-07
To answer your second question, I made a benchmark of the radix4 algorithm, with and without just the twiddle factor change.

Without:
```
test radix4_00000064 ... bench:         270 ns/iter (+/- 11)
test radix4_00000256 ... bench:       1,479 ns/iter (+/- 227)
test radix4_00001024 ... bench:       7,355 ns/iter (+/- 504)
test radix4_00004096 ... bench:      37,239 ns/iter (+/- 2,340)
test radix4_00016384 ... bench:     186,751 ns/iter (+/- 7,995)
test radix4_00065536 ... bench:   1,023,453 ns/iter (+/- 153,910)
test radix4_01048576 ... bench:  37,359,833 ns/iter (+/- 3,494,208)
test radix4_16777216 ... bench: 1,548,138,561 ns/iter (+/- 95,935,202)
```

With:
```
test radix4_00000064           ... bench:         266 ns/iter (+/- 90)
test radix4_00000256           ... bench:       1,423 ns/iter (+/- 75)
test radix4_00001024           ... bench:       7,154 ns/iter (+/- 303)
test radix4_00004096           ... bench:      35,719 ns/iter (+/- 1,741)
test radix4_00016384           ... bench:     176,436 ns/iter (+/- 9,357)
test radix4_00065536           ... bench:     840,508 ns/iter (+/- 79,022)
test radix4_01048576           ... bench:  23,791,329 ns/iter (+/- 912,981)
test radix4_16777216           ... bench: 555,863,689 ns/iter (+/- 24,350,419)
```

This is only the twiddle factor change, everything else is present in both. So the change has no effect when the whole problem fits into L2 cache (64, 256, 1024), but as it gets bigger, there's a more and more pronounced change.

So I was slightly wrong, it's not just 10-20%, it has a massive effect on large sizes because of cache performance. But a mixed radix step via the planner does an even better job of being cache friendly than this improvement to radix4, and so via the planner you don't notice the improvement to radix4 as much.

For the size 65536 specifically, you get a 20% speedup AND it's faster then planning a mixed radix.