@ejmahler 2017-08-07
I updated my comment on the memory usage. Before, I calculated it to be ~1.25n, but it's actually just n