@awelkie 2017-08-07
Thanks for the clarification. This is a great improvement!